refactor: move link templates
